Даниярдын акыркы озгорткону

+++ AdminKafe/ViewModels/CoolGet.cs
@@ -5,6 +5,6 @@
         public string FoodName { get; set; }
         public int FoodCount { get; set; }
         public double FoodPrice { get; set; }
-        public int Gram { get; set; }
+        public double Gram { get; set; }
     }
 }
